pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "idna==3.11",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.1.13",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.31.0",
    "pillow==12.2.0",
    "python-dotenv>=1.2.2",
    "requests==2.33.1",
    "tiktoken>=0.12.0",
    "urllib3==2.6.3",
    "uvicorn>=0.44.0",
]
